requirements.txt
cloudinary>=1.11,<1.12
Flask>=1.0,<1.1
Flask-Cors>=3.0,<3.1
PyJWT>=1.6,<1.7
uWSGI>=2.0,<2.1
